tutorials/github.com/alice52/awesome/test/redis/redis_op.go
package redis

import (
	"context"
	"time"

	"github.com/go-redis/redis/v8"
)

const (
	KeyValidWebsite = "app:valid:website:list"
)

func GetKey(rdb *redis.Client, key string) bool {
	// 这里可以是对redis操作的一些逻辑
	ctx := context.TODO()
	if !rdb.SIsMember(ctx, KeyValidWebsite, key).Val() {
		return false
	}
	val, err := rdb.Get(ctx, key).Result()
	if err != nil {
		return false
	}

	// 设置 blog key 五秒过期
	if err := rdb.Set(ctx, "blog", val, 5*time.Second).Err(); err != nil {
		return false
	}
	return true
}
